CareerItemコンポーネントを作成し、表示処理を更新
- components/ui/CareerItem.tsxを新規作成
- renderValue関数をCareerSectionからCareerItemに移動
- boolean型の値のみを表示するように実装
- UI/UXの一貫性を保持（true は ◯（緑色）、false は ✗（赤色））
